requirements.txt
uvicorn
fastapi
videosdk
openai
python-dotenv
elevenlabs
numpy
deepgram-sdk==3.4.0